Don't allow non-dialog/alert windows from unknown apps to be active window
This is the second attempt to deal with the following Gecko bug:
https://bugzilla.mozilla.org/show_bug.cgi?id=1882794

The first attempt (commit 85da96eff9b) had to be reverted due to side
effects with badly-behaved apps that flood us with events. Checking
the app in those cases is non-performant.

This attempt is more conservative and waits until we check to see if
the window can be the active window. Ideally, it would treat all
windows that come from apps unknown to AT-SPI2 as invalid because
that condition should not happen. However, if we do that, Orca will
not fully present dialogs and alerts from Gecko apps.

This approach seems to accomplish the goal of preventing unnecessary
keygrab refreshes, title repetition, and bogus focus updates without
making Thunderbird super laggy. Dialogs and alerts are also being
presented.