soffice: Use accessible ID to detect spellcheck dialog
For LibreOffice >= 25.2, the spell check dialog has
an accessible ID of "SpellingDialog" set, so use
that to reliably detect it. Only fall back to
using the previous heuristics when no accessible
ID is set for the dialog being observed (still needed
for LibreOffice <= 24.8).

For more details, see the discussion in the corresponding
LibreOffice Bugzilla ticket tdf#155447 [1] and
the LibreOffice commits from which on the accessible ID
is reported for both, the gtk3 and qt6 UI variants [2] [3]:

    commit 8495e77f4d49733f645cdacfb71729b91e5c4f30
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Jul 16 15:03:20 2024 +0200

        tdf#155447 gtk3 a11y: Use GtkBuilder ID as accessible ID

    commit c64b823b74cbd3063d6e87643bd68ea5343b2f54
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Jul 16 13:57:41 2024 +0200

        tdf#155447 qt a11y: Report accessible ID

The approach could be extended to also use identify
more individual relevant objects in the dialog
by their ID.

[1] https://bugs.documentfoundation.org/show_bug.cgi?id=155447
[2] https://git.libreoffice.org/core/commit/8495e77f4d49733f645cdacfb71729b91e5c4f30
[3] https://git.libreoffice.org/core/commit/c64b823b74cbd3063d6e87643bd68ea5343b2f54