Fix braille for multiline text areas
In commit 4bcbb868bf20, _generateContext() was no longer getting called;
we were simply checking the keyword arguments for "includeContext"
in _generate_default_prefix(). Move the logic from _generateContext()
into _generate_default_prefix(). This will cause us to stop generating
ancestors for multiline text fields when we're not on the first line.
That was the behavior prior to 4bcbb868bf20.

Note that we should be able to generate ancestors even in the case of
multiline text areas. However, the way braille.getLineInfo() splits
the line into words can result in incorrectly grouping the last word
of an ancestor with the first word of the line if that line is empty
or has leading whitespace characters. For now, mark the code in
question with a TODO.

Lastly, remove the undocumented, GUI-less setting "enableBrailleContext".
We should be smart about when to generate ancestors and ensure the
display starts with the relevant information.