Ignore selection-change event when current focus is among selected children
When selected multiple messages in Thunderbird, a focus-change event is
received first followed by the selection change. We were presenting both,
but the former was getting cut off by the latter. And the latter currently
treats the first valid, non-layout-only selected child as the item to
present due to selection in table cells where there are no accessible row
objects. That issue will be addressed separately. To solve the immediate
issue, it should be safe to ignore the selection-change event when the
currently-focused object is among the selected children.