Web: Adjust presentation of loading-related messages
Commit 3fa2e7d9095 completely disabled all messages related to loading
of page content if the user's verbosity was set to brief. This was
done in response to a user request. However it seems a bit extreme
because in many cases absolutely nothing will be spoken to indicate
that the content is ready to be read.

In order to balance things a bit, adjust what we present if the
verbosity is brief:

1. No loading start messages will be presented.
2. "Finished loading" will not be presented.
3. Only the name of the newly-loaded page will be presented.

Hopefully this will address everyone's needs without the need for even
more settings.