Move management of Orca modifier pressed state to manager
We were managing the pressed state of the Orca modifier in the
KeyboardEvent class and updating it based on new events. That is no
longer reliable in the new-keyhandling world. Given the following
event ordering:

1. Orca modifier pressed event received
2. window:deactivate event received
3. Orca modifier released event received

We set the state to pressed. Then we remove modifier grabs if the
window:deactivate event causes script deactivation. Then when we
get the release event, we conclude it's not for the Orca modifier
because we have no grabs for that key. As a result of this sequence
we never update the pressed state of the Orca modifier until it is
released without anything triggering the removal of modifier grabs.

To clean this up a bit, move management of the pressed state to
the OrcaModifierManager so that it can clear the state prior to
removing modifier grabs.