Event Manager: Don't change the active script for unknown apps
This is in response to https://bugzilla.mozilla.org/show_bug.cgi?id=1882794.
When we get events from apps that are not registered with AT-SPI2, don't
update the active script. Doing so causes unnecessary key grab refreshing
and other clean up related to changing scripts. It may also be the source
of a bug reported by a user which I cannot reproduce, that he started
experiencing after the old workarounds for the above bug were removed.