Fix SayAll failing when launched from non-text web element
The web script used to have its own say-all method that was redundant
to the one in the default script and thus removed. However, the default
script's method returned early if SayAll was launched from an object
which had no text or only whitespace. As a result, launching SayAll
from an image or separator caused Orca to only present that item.

Removing the aforementioned check fixes the web-content regression
and shouldn't impact other applications: in the case of empty/
whitespace-only objects, we'll either move on to what follows and
present it or not find anything that follows and return.