Eliminate use of displayedText()
This function is used quite a bit to get the text of an object which
implements the accessible text interface. But it falls back on -- or
sometimes prefers -- the accessible name. Individual callers should
ask for the text when they want text, the name when they want the name,
do any fallback work needed, etc.

See issue #528.