Web: Always handle sayCharacter/Word/Line() for document content
We were deferring to the default script's sayCharacter/Word/Line()
if the element was not editable and our caret navigation was not
being used. As a result, if one used the browser's native caret
navigation (i.e. F7 + switching to focus mode), any embedded object
in the character/word/line would not be expanded and the user would
hear however their synthesizer handles U+FFFC (often it's not spoken).

We should do our best to present the same character/word/line regardless
of whether we're in browse mode or focus mode, therefore only defer to
the default script if we're not in document content.

Note: There is an additional issue not addressed by this commit, namely
if native browser navigation in focus mode causes a link to become
focused, we only present that object rather than the unit of text
associated with the navigation. That will be addressed in a follow-up
commit.