Web: Speak the correct unit of text regardless of mode
locus_of_focus_changed determines what unit of text should be spoken
based on various and sundry conditions (object event, type of object,
input event, etc.). However, it wasn't handling changes in focus related
to caret navigation in focus mode. This commit causes navigation to be
treated the same regardless of mode. It also eliminates a number of
special cases that should be obsoleted as a result.