Web: Fix link regression better
The removal of displayedText() and all of its associated messiness
in c08c49b9e2c regressed link presentation when the link had one or
more children. 25224a6f6a fixed that for links with text children
but not for children of other types, such as image links.

Looking further into it, it turns out that the web script's
_generateName() was returning nothing for links which were not
given an explicit name. Then displayedText() ignored the link text
due to the presence of "U+FFFC" and returned the name as the
displayed text. Thus the real fix is to have _generateName() do
what it claims to do.