Web: Don't present the role name of an editable element if the object hasn't changed
Starting with 164ce522c0, sayCharacter/Word/Line() is always handled by
the web script for document content. As a side effect of this change,
Orca started speaking "editor" before each VSCode line.

There is already logic in _generateRoleName() which checks if we're in
the same editable element and suppresses the role name if we are. But
the logic for aria-roledescription comes first and if it finds an
author-provided role, presents it. The VSCode editor has such a role.

Fix this regression by adding a check for this condition prior to the
aria-roledescription logic.