Revert "script-utilities: Present unrelated labels for list items"
This reverts commit 50a45a45. That commit restored list items to objects
for which we'll do a tree dive in a desperate attempt to find something
to present in response to a focus/selection change, and was made in order
to restore presentation of Fractal room history.

Since then, the code which made that desperate attempt (displayedText())
has since been removed (see commit c08c49b9). As a result, the motivation
for the change in commit 50a45a45 is obsolete.