Temporarily special-case editable content in web sayLine()
Commit 164ce522c caused the web script to always handle speaking the
character/word/line for document content. As a side effect, Orca stopped
speaking indentation in VSCode. The reason why is that the adjustments
to the string to be spoken are not currently being made in the generators
as they should; instead they're being made on the spot in the default
script's sayLine().

Given that doing the work where it belongs may be a non-trivial change,
let the default script's sayLine() handle presentation of a line of text
which lacks U+FFFC and is in an editable object for now.