Work around GTK2 (and possibly 3) brokenness
When expanding a submenu in (at least) Pidgin, and rumor has it Geany,
the newly-expanded menu lacks both the showing state and the visible
state causing us to ignore any object:selection-changed events it fires.
That wasn't a problem before because we were unconditionally presenting
anything that claimed focus via the deprecated-over-a-decade-ago "focus:"
event. We no longer listen for that event and have thus been bitten.
Work around it for now so the user experience doesn't regress.