Use libatspi, not libatk to get localized role name
Use `atspi_role_get_localized_name` [1] from libatspi
directly to get a localized name for an `Atspi.Role`,
instead of first trying to convert that to an AtkRole
and then using `atk_role_get_localized_name` [2] from libatk.

This is simpler, and the previous assumption that ATK and
AT-SPI role names are the same is at least currently not
true for (push) buttons after at-spi2-core commit [3]

    commit 0c2ebfad9ebf92913a326143345b841d54fa1866
    Date:   Tue Aug 6 16:02:55 2024 -0500

        Rename ATSPI_ROLE_PUSH_BUTTON to ATSPI_ROLE_BUTTON

, so Orca would announce "mybutton, invalid" instead
of "mybutton, button" or "mybutton, push button" when
a (push) button receives focus.

With this commit in place, it now says "mybutton, button"
when using a libatspi version containing the above commit,
which is the intended behavior as described in [4].

Note: This additionally needs pending at-spi2-core MR [5] in place
for Orca to work at all after the above-mentioned
at-spi2-core commit.

[1] https://docs.gtk.org/atspi2/type_func.Role.get_localized_name.html
[2] https://docs.gtk.org/atk/type_func.Role.get_localized_name.html
[3] https://gitlab.gnome.org/GNOME/at-spi2-core/-/commit/0c2ebfad9ebf92913a326143345b841d54fa1866
[4] https://gitlab.gnome.org/GNOME/at-spi2-core/-/issues/173
[5] https://gitlab.gnome.org/GNOME/at-spi2-core/-/merge_requests/166