Also set the prior object for structural navigation and caret navigation
The web script's presentObject() was setting the prior object in the
case of caret navigation to eliminate chattiness. We should also do
this in the case of structural navigation and table navigation. This
is also a speculative fix for one of the issues described in issue #532.

Looking at the related structural navigation code, things can be made
much simpler and less redundant. That was the case already, but is
even more so after this commit. Cleaning up that logic will be done in
a follow-up change.