Move setting caret position out of structural navigation
In order to set the caret position for web content, we first do the work
to figure out the first caret location. This work is then essentially
duplicated when determining the line/object contents. Therefore, remove
the calls from structural navigation and set the caret position to the
first presentable object determined by sayLine()/presentObject().