Improve name-description redundancy check
We use SequenceMatcher to determine if a string (e.g. the description)
is redundant to another (e.g. the name). It turns out that if one is
really long and the other is short but fully contained in the other,
SequenceMatcher returns a low threshold. And a check for that case.