Don't present the full line when arrowing to the end of an input element
When we generate the presentation of an input element, _generate_text_line()
defers to _generate_text_substring() in case we need to handle something like
caret navigation. But _generate_text_substring() was only returning something
if the string was not empty and also lacked embedded object characters. In
the case of navigating to the end of an input element, the string will be
empty. It should still be presented. Therefore only return [] if there's an
embedded object character in the string.