Clear the last non-modifier event after modifier press+release
We have event-processing logic based on the last non-modifier input event.
For instance, if the last input event was Down and a caret-moved event
comes in, we should present the line of text just arrowed to. However, we
were not clearing this in the case that the last input event was a press
and release of a modifier such as Ctrl. We should have been in order to
prevent presenting the wrong thing.