Web: Don't present line for newly-focused element reached via character
edb0d02466 fixed the failure to announce new blockquotes in editable
messages when navigating by line. As a side effect, it caused us to
announce new paragraphs and sections reached via character nav instead
of just announcing the current character. Fix that by letting the default
script handle presentation when navigation is not by line and the content
is editable.