Fix "chattering" when selecting text in Thunderbird
It turns out that 7e2169fadf, which fixed a previously-unnoticed use
of pyatspi, unmasked a bug in which Orca is extremely chatty when
selecting text in focus mode. That chattiness did not occur before
because an exception was thrown. To eliminate the immediate problem,
have the default script process text selection changes when focus mode
is active. This causes Orca to behave as it did in version 46.

Note that Orca was not, and still is not, presenting the selection
correctly in editable, rich-text messages. This will be addressed in
a follow-up change.