Prefer the text line over the name for brailling a list item
Editable lists items in Thunderbird and browsers are unfortunately
being given an accessible name based on the contents of that list.
As a result, if we prefer the item's name over its text in braille,
it will appear that the cursor is not being updated. Prefer the
text in this case and fall back on the name.