Mouse Review: Remove sanity check for active document
Having removed all the workarounds for a mozilla bug [1], sanity
checking the active document sometimes fails, in which case we don't
present anything. If we have a valid object under the pointer, treat
it as if it came from the active document. If that proves problematic,
we can special case that scenario.

[1] https://bugzilla.mozilla.org/show_bug.cgi?id=1882794