Hack around Qt giving us a new cell every time we ask for a child
a6116995bff6 broke presentation of Qt cells in QTreeView. That change
eliminated some chattiness by not presenting the name of an object
that was not the focused object if the name was identical. It turns
out that when we go to generate the name for one of these cells, the
equality check fails. We get a new copy of the object every time we
ask for the child via index.

As a safe hackaround in time for the release candidate, only do
the chattiness reduction if obj is not focus and obj and focus have
different roles. That should solve chattiness in the case of popups
with the same name as the focused object, as well as the redundant
labelling we're seeing in GTK4 (libadwaita) apps.

See issue #535.