Eliminate more roles from _generate_descendants()
For GUI list items, we need to be able to handle interactive descendants.
But the Thunderbird message list items have named sections as descendants
causing us to be chatty. Therefore add roles that are common to web content
but not GUI list items to those we ignore when generating descendants.