Partially revert ac3239d483; fix link chattiness slightly differently
ac3239d483 removed special casing of links from _generate_accessible_name.
That special casing had been in place to make link presentation less chatty,
but it caused presentation of links in the side bar of GMail to break. While
ac3239d483 fixed links in GMail's side bar, it reintroduced chattiness in
other places, such as when a link was split into multiple objects due to
having a child object.

This commit restores the special-casing, but instead of checking if the
link has presentable text, it instead checks to see if we've already
figured out the substring we want to present. If we've already figured
out the substring, use it rather than the name. This solution appears
to work in both scenarios.