Include description in the name chattiness filtering
_generate_accessible_label_and_name() attempts to eliminate chattiness
with labels, ancestors, etc. By filtering the name if the object isn't
the focused object and the name of object is either the name of the
focused object or the description of the focused object. This does
not catch the following scenario: The focused object has a name,
an ancestor object lacks a name BUT has a description which is the
same as the name of the focused object. When generating the name of
the ancestor, we'll fall back on the description thus presenting it
twice.

Add a similar fallback in _generate_accessible_label_and_name(),
so if the name/description is the name or description of the focused
object, filter it out.