Eliminate more chattiness from redundant descriptions
_generate_accessible_description was not handling the following
scenario: object is an ancestor of the current focus and the
description of the object is the name or description of the current
focus. When that occurs, we'll present the description for the
ancestor and then as the name or description of the newly-focused
object. Add a check for this condition to eliminate the double
presentation.