Event Manager: Replace flood handling with flood protection
Instead of detecting event floods from apps and doing our best to
mop up the mess, just keep track of the last app and time that
fired a given event type. If we get an incoming event of the same
type by the same app within 0.1 seconds, ignore that event if
the object emitting the event is not the current focus, is not
focused, and is not selected. This seems to catch more floods
and improve performance with apps that insist on firing way too
many events for changes to the accessibility tree.