Clear cache when a page tab becomes selected but lacks showing state
Newly-added mate-terminal page tabs sometimes lack the "showing" state.
As a result, if the user creates a new terminal, adds some text, and
attempts flat review, we'll ignore the terminal content because the
ancestor page tab claims to not be showing.

Try clearing the cache to see if the problem is just a missing state
change event.