Braille: Don't kill flash message in default.process_routing_key()
default.process_routing_key() calls braille.process_routing_key().
The latter, upon finding a flash message, restores the pre-message
contents and returns early. But if it doesn't find a flash message,
it routes the cursor to the specified location, possibly triggering
a click action. As a result, when default.process_routing_key()
kills the flash message first, the user thinks they're clearing the
flash message, but they may instead be clicking on something. Fix
this by interrupting presentation with killFlash set to False.