Fix double execution for a newly-bound command
If a command was bound in the Orca preferences dialog and then that
newly-bound command was used without restarting Orca, that command
would be executed twice. This is due to the event manager being
reactivated without having first been deactivated in loadUserSettings()
and a second keywatcher being created.

While we could deactivate the event manager in loadUserSettings(),
loading user settings should not have anything to do with object events
or input events. Therefore move initial activation of the event manager
into main() and pause events in loadUserSettings().

In a similar spirit, it's unclear why the script manager should be
deactivated and reactivated due to loading user settings. Moreover,
doing so causes script event listeners to be unregistered and re-
registered which seems unnecessary. Therefore move activation of the
script manager to main().

Lastly, check and handle double activation/deactivation of the event
and script managers which should not happen.