Gecko: Ignore state-changed:focused events for the frame
We're sometimes seeing state-changed:focused events for (at least) the
Firefox frame. This can cause us to update the locus of focus to not be
in document content, at which point browse mode commands fail. Address
this by ignoring this event when it comes from the frame. We should (and
do) get window events for frame activation.