Clear the cache when we get a focused event for an object which lacks focused state
When an app fires a focused state-changed event with detail1 = True, but
the object lacks the focused state, we were returning early assuming that
a subsequent event caused the object to lose focus. Instead, it could just
be an app bug. And if we trust the state set, we might fail to present the
event. Therefore clear the AT-SPI2 cache non-recursively and check again
before returning early.