Script Manager: Call reclaim scripts less often
We check for apps which are not in the desktop as part of cleaning
up scripts. Doing this work for every window-related event is unnecessary
and non-performant. The previous commit ensures we clean up scripts when
we are notified by the desktop frame of child removal. That, plus in
response to a window:destroy event should be enough.