Don't swallow exception if we cannot connect to the a11y applications gsetting
While Orca is a screen reader for all desktop environments, there are
enough GNOME dependencies that orca.gnome.desktop.a11y.applications should
not throw an exception when we attempt to access it via Gio.Settings. If
that's not the case, I'd like to understand why and handle it explicitly.
Therefore print a "severe" message if this fails.