Make the timeout handling a non setting
Timeout handling as a setting was done early on in Orca's creation,
possibly as a shortcut for the original developers. It should not be a
setting. There is now only one place in which we handle a timeout, namely
in orca.shutdown(). Therefore remove the associated settings. Also reduce
the timeout time from 10 seconds to 5. Shutdown should be immediate. It
should not reach 5 seconds, let alone 10.