Remove crashOnSignal(); adjust shutdownOnSignal() to handle crashes
Handling SIGSEGV is generally not recommended. But we try in order to
restore CapsLock in the event of a crash. Whether or not we should
continue this remains to be determined. In the meantime, adjust
shutdownOnSignal() to prioritize restoring CapsLock and to exit without
further cleanup in response to a crash. Remove crashOnSignal().