Work around https://bugreports.qt.io/browse/QTBUG-130116
If the active window is being set and our stored focus is not in that
window, we update the current focus to the active window. Otherwise
we are in danger of basing decisions using the wrong object/script/etc.

Determining if our stored focus is in the current window assumes that
we can ascend the accessibility tree from the object to its window.
When that fails, e.g. due to https://bugreports.qt.io/browse/QTBUG-130116,
the conclusion would be that focus is not in the active window and we'd
update the focus to the active window and present it.

Work around the Qt bug to check for broken ancestry for all Qt objects.
If the ancestry is broken, don't update focus to the active window
and hope for the best.