Don't speak name of non-focusable section unless we know it's author specified
Chromium-based browsers appear to be applying name-calculation to certain
divs which are not focusable. This causes us to double-present information,
namely speaking the div name (which came from its contents) followed by the
current line (i.e. the contents).

When the author explicitly provides a name, e.g. via aria-label/labelledby,
browsers indicate this via object attribute. If that attribute is absent and
the section is not focusable, treat this as a generic text object and do not
include the name.