Don't speak name of non-focusable section
Chromium-based browsers appear to be applying name-calculation to certain
divs which are not focusable. This causes us to double-present information,
namely speaking the div name (which came from its contents) followed by the
current line (i.e. the contents). Therefore if the section is not focusable,
don't include its name.