Web: Set default value of `priorObj` to `old_focus` in locus_of_focus_changed()
When `priorObj` is not explicitly set, the current focus is used. This makes
sense when we're about to present something that will become the focused
object. However, if we had to update focus prior to presentation, this can
cause us to conclude incorrectly that focus didn't actually change. And we
we think focus didn't change (e.g. arrowing from one line of a heading to
another line in that same heading), we try to be less chatty.

locus_of_focus_changed() was sometimes setting `priorObj` to `old_focus`.
That should be the default, and we should special-case (and document) any
instances where doing so is not desired.

See issue #541.