Limit speech generation for ancestor sections
Most sections are excluded from ancestor generation because they are
deemed layout-only. But if an ancestor section fails the is-layout-only
test (e.g. due to having a name or being focusable), we will attempt to
generate speech for it. This can result in some content being presented
twice. Therefore, when generating speech for an ancestor section, limit
the generation to the bare minimum and return before we do any text-related
generation.