Try to filter out redundant description-change events
Based on some debugging output from a user. Some (unreleased?) version
of Dolphin does the following each time selection changes:

1. Fires the focused-changed event for the newly-selected item causing
   Orca to announce that item's name and description.
2. Fires a description-changed event for the list, with the description
   being set to the name + the description of the item we just announced.

Starting with Orca v47, Orca presents description changes not only on
the focused object but on its ancestors. This makes it possible for app
developers to associate a description with a group of widgets and update
that description as needed. Unfortunately, it means that we now double-
speak every selected item in whatever version of dolphin has this change.

In order to try to filter this sort of thing out, apply the string
redundancy check comparing the ancestor description to the combined name
plus description of the current focus. If they are simmilar enough, do
not announce the description change.

See issue #542.