Try to eliminate some chattiness in listitems with children
Because GTK4 lists seem to be used everywhere, and can contain multiple
descendants which app developers expect Orca to present automatically,
Orca v47 and later now looks for and speaks descendants with presentable
information.

Unfortunately, there are apps who decided to address this need themselves
prior to Orca v47, by cramming all the info into the accessible name of
the list item. As a result Orca v47 and later are sometimes super chatty
because Orca presents the name and all the descendants which have something
presentable.

In an attempt to maintain reasonable presentation of both types of app
implementations, make the following changes when generating descendant
presentation:

1. If the item lacks a name, fall back on the labelled-by text.
   There's not a real-word case of this yet. But it's consistent
   with the next item.
2. If the item lacks a description, fall back on the described-by text.
   It doesn't appear that described-by text is automatically exposed
   as the description.
3. When determining if a label's name is redundant to the object name,
   use the string-redundancy check. This is needed because some app
   developers are not using the exact descendant name in the item name.
   A little bit of string massaging is enough to make the existing
   filtering which landed in Orca v47 fail.
4. When determining if a label's name is redundant to the object
   description, also use the string-redundancy check for the same
   reason.
5. If the label's name is redundant to the item description, don't filter
   it out as the descendant; instead don't present the description of the
   object. Example: In Tuba, the toot being exposed as the description
   means Orca will present everything else about the item with the toot
   being last. Instead we want to present the toot in the correct location
   with respect to the other descendants, namely above the status buttons.
6. If the full name of a button is completely contained within the ancestor
   item name, don't present that button. This was also Tuba inspired (the
   status buttons at the end have just a number as their name, and those
   numbers are summarized in the name).

Note: Excluding an item like a button (or anything else) during the
presentation of the item name has nothing to do with presenting that
item when it gains focus (e.g. via Tab within the item).