LO: Adjust workaround for bug 161444
In versions of Calc where the bug is fixed, we get the desired focus
change for the table (yay!). But the on_active_descendant_changed
workaround kicks in if the table firing the event is not an ancestor
of the current focus. In the fixed version of Calc, the table firing
the event is the current focus, so now we're presenting the sheet
twice.

Fix this by adjusting the workaround to not kick in if the table is
the focus. This prevents double-presentation in fixed versions while
keeping the logic in placed for older versions.