LibreOffice: Ignore focus claims from ancestor root pane of current focus
After the Writer/Calc window becomes active, the current object in the
document claims focus (good). But immediately after an ancestor root pane
claims focus even though the user is still in the document object. There
does not seem to be any reason to present this focus change. It's just
unnecessary chattiness. Furthermore, it is sometimes causing speech to
be interrupted before the real focus is fully announced.