LO: Follow-up change needed to ensure we announce sheet changes
5e573ec9c2 and 153b255 were needed to ensure we presented the newly-focused
sheet and did so only once when the Calc window was activated. But the sheet
was no longer being presented when the sheet was changed. This was because
a heuristic to detect post-editing focus claims was kicking in and updating
focus silently.

Fix that by adjusting the heuristic to check that the table claiming focus
is an ancestor of the current focus. If it's not, we're likely changing
sheets. This fix alone, however, causes the frame to be announced in
addition to the sheet. So set the locus of focus to the parent of the
table silently prior to the default script handling the focus change.