Don't attempt to calculate accurate posinset and setsize in giant sets
We normally attempt to calculate an accurate posinset and setsize for
items in a collection when those values have not been provided to us,
e.g. via aria-posinset and aria-setsize (or the native app equivalent).
But before we do that work, we check check the member-of relation, the
node-parent-of relation, do some special handling for description lists,
tables, menus, etc. In the end, however,  we sometimes wind up iterating
through all of the children to determine the membership.

This iteration is no big deal until the set size gets large (say, over 1000
members) at which point we become laggy. Given a set size in the tens of
thousands, we grind to a halt. To see this halt in action, create a folder
with 10,000 items and then use a file explorer to navigate among the items.

In order to be performant, if the parent container has more than 500
children, return its child count as the set size and the item's index
in parent as the position in set. In cases where this result is not
accurate, the app in question should expose the correct position in set
via the "posinset" object attribute on the item and the set size as the
"setsize" object attribute on the same item.