Live regions: Clear last message in response to flushMessages()
If the same live region message results from two different back-to-back
key presses, we can incorrectly conclude that the second message is a
duplicate that should not be presented. In order to prevent that from
happening, clear the last message in response to flushMessages() which
gets called as part of the presentation interruption we do for key
presses.