Web: presentObject() should defer to the default script's for alert presentation
As described in https://bugzilla.mozilla.org/show_bug.cgi?id=1895857,
Orca is super chatty when presenting certain alerts when aria-labelledby
and aria-describedby are used and the label/description point to objects
in the alert. We should add filtering to getObjectContentsAtOffset() for
this condition. However, an ARIA alert should be presented as UI; not
as a text container whose contents should be pieced together. Deferring
to the default script's presentObject() accomplishes this and eliminates
the chattiness.