Have AXObject.get_role() return PANEL when we get GROUPING
For all ARIA roles with the expected mapping of PANEL, GTK4 instead gives
us GROUPING. This is not a problem in Orca v48 where there is already
logic that explicitly handles the GROUPING role. However, backporting the
necessary logic in v47 is a bigger change than I would like to make for a
stable version. Therefore, modify AXObject.get_role() to return a PANEL
instead of GROUPING.

See issue #546.