Fix lint and other nits in spellcheck support
Also move common/shared implementations from the subclasses to the
superclass and update some of the heuristics to match changes in
the supported applications.

Also create AXObject.get_accessible_id() and use it in the LO spellcheck
code because there should not be any direct calls to AT-SPI2 functionality
outside of the AX* classes.

See issue #490