Spellcheck: Check for accessible id in more places
Do the following case-insensitive checks:
* If the object's accessible id starts with "suggestions" treat it
  as the suggestions list.
* If the object's accessible id starts with "replacement" treat it
  as the object (likely entry) which contains the proposed replacement.
* If the label's/widget's accessible id starts with "error" treat
  it as the container displaying the misspelled word.

Note that the first of the three is based on what LO 25.2 currently
exposes ("suggestionslb"). The other two are not in use yet, but adding
them facilitates implementation in, and getting feedback from, apps and
toolkits.

Also modify the existing check for the window. We were doing an exact
match on "SpellingDialog". Making that case insensitive and limiting to
starts with "spelling" works with the current LO implementation and
removes an implementation detail ("dialog").