Web: Simplify logic related to non-element and pseudo-element objects
We were identifying list item markers, CSS generated content nodes,
and non-generated text nodes separately. In all cases, however, we
want to ignore them because the presentable content is in the accessible
text of the parent element. Therefore create a new is_web_element()
utility and use it to identify all of these ignorable nodes.