Remove check and workaround for non-showing objects in on_caret_moved()
We normally don't want to present caret-moved events from objects which
are not showing. Therefore, on_caret_moved() checked to see if the source
of the event was showing. If it was not showing, it would return early
unless the event came from the current focus. Subsequent code returns
early if the event is not for the current focus. Therefore there's no
benefit gained from checking if the event came from a showing source,
other than debugging info (which is included already when processing the
event).