AXUtilitiesEvent: Accept checked-changes from descendant of interactive item
If focus is on a list item or tree item, and the user presses space
causing the checked state of a descendant to change, we should announce
that change even if the checkable item itself is not focused. There was
already logic in the GTK script for this condition, but it can apply to
other toolkits, web apps, etc. Therefore move the logic to AXUtilitiesEvent.