Ignore text-selection-changed events whose reason is unknown
If we don't have a clue about why the text selection changed, ignore
the event to eliminate some chattiness. Given that we check for many
possible reasons in _get_text_selection_changed_event_reason(), this
change will hopefully not eliminate presentation that we should do.
But if it does, that means we need to determine and add a new
TextEventReason to prevent that presentable change from being ignored.